refactor(layout): Update Ticket Layout jsx -> tsx including finish PatternFly conversion
ref: #209 #205 #146
